### 2023-04-29 #5426 @arfon
`cloc` output for default (`master`) branch:

```
cloc TDLM/
      41 text files.
      22 unique files.
      36 files ignored.

Language                     files          blank        comment           code
R                               12            183            972           2455
Java                             3            132            153            889
Rmd                              1            100            296            126
TeX                              1             12              0            112
YAML                             3             20              4             88
Markdown                         2             17              2             25
SUM:                            22            464           1427           3695
```
