### 2023-05-03 #5434 @editorialbot
[REVIEW]: TDLM: An R package for a systematic comparison of trip distribution laws and models
**Submitting author:** @maximelenormand (<a href="http://orcid.org/0000-0001-6362-3473">Maxime Lenormand</a>)
**Repository:** https://github.com/EpiVec/TDLM
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.1.1
