### 2023-05-10 6a38e15
Optimize distance extraction in extract_spatial_information()
