### 2023-05-10 2f2c7ef
Update extract_spatial_information, NEWS and README
