+++ NEWS.md
@@ -1,4 +1,9 @@
-# TDLM v0.1.0 
+# TDLM (development/github version)
+
+* `extract_spatial_information()` is now optimized and way more efficient 
+regarding the computation of distances (thanks to @Pachka).
+
+# TDLM 0.1.0 
 
 First release on CRAN
 
