+++ README.md
@@ -5,7 +5,6 @@
 <!-- badges: start -->
 [![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/TDLM)](https://cran.r-project.org/package=TDLM)
 [![rstudio mirror downloads](http://cranlogs.r-pkg.org/badges/grand-total/TDLM)](http://r-pkg.org/pkg/TDLM)
-[![version](https://img.shields.io/github/v/release/EpiVec/TDLM?label=version&style=flat&logo=github)](https://github.com/EpiVec/TDLM) 
 [![licence](https://img.shields.io/badge/Licence-GPL--3-blue.svg)](https://www.r-project.org/Licenses/GPL-3) 
 <!-- badges: end -->
 
