+++ NEWS.md
@@ -1,7 +1,12 @@
-# TDLM (development/github version)
+# TDLM 0.1.0 (development/github version)
+
+This is a list of changes made in the development/github version of the package
+between TDLM 0.1.0 (CRAN release 2023-03-18) and the next CRAN release.
+
+* Fixed bug in `extract_spatial_information()` (@Pachka, #1).
 
 * `extract_spatial_information()` is now optimized and way more efficient 
-regarding the computation of distances (thanks to @Pachka).
+regarding the computation of distances (idea proposed by @Pachka).
 
 # TDLM 0.1.0 
 
