### 2023-06-04 EpiVec/TDLM#3 @kanishkan91
JOSS review: No test cases
There are no test cases available for this package. This is a prerequisite for scientific software and publication in JOSS. Test cases ensure that future development will be tested and stable. I would recommend the author does two things,

1. Add test cases - As an example I add a link to the test cases folder in a another package- https://github.com/JGCRI/gcamdata/tree/main/tests
2. Add a code coverage badge that will show how many lines of code in every file are covered by a test case.

Ideally atleast 60% of your code should be covered by a test. Let me know if you have any questions related to this.

https://github.com/openjournals/joss-reviews/issues/5434
