### 2023-06-04 EpiVec/TDLM#4 @kanishkan91
JOSS review: Recommended edits to README
1. I would recommend adding a short summary of what the package does in the README. You can use what you wrote for the paper directly.
2. Could the author also add a "Contribution instructions" section so that the users would know who to reach out to in case of issues. Author would just need to mention an email ID and encourage users to open issues.

https://github.com/openjournals/joss-reviews/issues/5434
