### 2023-06-04 EpiVec/TDLM#5 @kanishkan91
JOSS review: No github CI  badge, windows, ubuntu builds? 
Can the author add the Github CI badge to the repo? It seems the workflow has been set up. But there is no badge on the repo. Should be simple to add this in.

Also, where is the build info for a windows or ubuntu build? Has that been set up?

This would ensure long term stable development of this package.

https://github.com/openjournals/joss-reviews/issues/5434
