### 2023-06-04 EpiVec/TDLM#6 @kanishkan91
JOSS review: What is the purpose of the java folder? Can it be deleted? 
There is this java folder in the main root folder of the package - https://github.com/EpiVec/TDLM/tree/master/java

And I saw it again in the isnt folder https://github.com/EpiVec/TDLM/tree/master/inst/java

Maybe I am missing something but could the author clarify what this folder is ? If not used , can it be deleted?

https://github.com/openjournals/joss-reviews/issues/5434
