### 2023-06-04 EpiVec/TDLM#7 @kanishkan91
JOSS review:  (OPTIONAL) Given high number of package dependencies, can a virtual env be created for this package (like renv)
The packages required by TDLM seem to require a high number of dependencies. When installing the development version of the package, I was asked to update several of my packages. This made the installation take much longer than expected.

It seems the package is sensitive to version numbers of dependencies. This may inhibit future development from multiple users.

A simple solution for this is just setting up a virtual environment that takes care of all dependencies. This can be implemented atleast for the development version. Here is an example of how renv can be set up - https://docs.posit.co/ide/user/ide/guide/environments/r/renv.html

Of course, if this is too much of an hastle at the moment, can be implemented in the future.

https://github.com/openjournals/joss-reviews/issues/5434
