### 2023-06-04 EpiVec/TDLM#8 @kanishkan91
JOSS review : package version for dependencies? 
The dependencies listed here - https://github.com/EpiVec/TDLM/blob/master/DESCRIPTION, don't have an associated pacakge number. Can the author specify what those are? I suspect this is pushing the installation time up significantly since the development version seems to install the latest of every package.

https://github.com/openjournals/joss-reviews/issues/5434
