### 2023-06-04 EpiVec/TDLM#9 @kanishkan91
JOSS review: Unbale to install package due to version of spatstat.explore 
This also relates to another issue I just opened. I am working on R version 4.04. When I tried to install the package , there was an issue with the spatstat.explore package installation.

I think this is because the DESCRIPTION file does not impose a restriction on any of the dependencies so R just tries to download whatever version (usually the latest). This leads to the package installation failing. This can be fixed by imposing restrictions on the dependencies. This can be done by editing this file -https://github.com/EpiVec/TDLM/blob/master/DESCRIPTION

https://github.com/openjournals/joss-reviews/issues/5434

![image](https://github.com/EpiVec/TDLM/assets/37234086/3e395105-9393-46db-b3af-b132d389366b)
