### 2023-06-04 #5434 @kanishkan91
@crvernon @maximelenormand. I'm going through my review and have been opening issues that the author can start addressing. However, I would like to raise a couple of points here sonner rather than later.

**1. Lack of test cases**- I could not find any test cases for this package. I think this is one of the requirements prior to publication in JOSS. I would recommend the author adds test cases and ensures atleast 60% of the lines of code are covered by a test . This is referenced in this issue- https://github.com/EpiVec/TDLM/issues/3
**2. Problems installing**- I had a few problems installing this package. This is referenced here- https://github.com/EpiVec/TDLM/issues/9. I traced this issue to a simple fix, which is that the DESCRIPTION file does not impose any version constrains on the dependencies. So, even though I am working with the correct version of R (greater than 4.0.0), the installation seems to fail since the package seems to want the latest for all packages (some of which require a higher version of R). This is referenced in another issue- https://github.com/EpiVec/TDLM/issues/8
**3. No windows/Ubuntu build on workflows**- Maybe I am missing something but I could not see a workflow for a windows build or Ubuntu build here. Is that correct? If so, those should be added.

I have opened other issues as well, but these are some of the bigger ones that may take some time addressing. Hence wanted to highlight right away. Please let me know if I missed anything in the above.
