### 2023-06-05 EpiVec/TDLM#10 @kanishkan91
JOSS review: Building vignettes fails when installing from GitHub
When I tried devtools::install_github("EpiVec/TDLM", build_vignettes = TRUE) with the build_vignettes set to TRUE, installation failed.

However, if that is set to FALSE, the package installs fine. Adding the error below-

![image](https://github.com/EpiVec/TDLM/assets/37234086/35487e69-7a75-4430-ba05-885739bb1b83)

https://github.com/openjournals/joss-reviews/issues/5434
