### 2023-06-05 f3f4f6f
Update README (#4), DESCRIPTION (#10) and vignettes (#11)
