### 2023-06-05 EpiVec/TDLM#4 @maximelenormand
I added a "Short description" and "Contribution instructions" in the README. Thank you for the suggestions.
