### 2023-06-05 EpiVec/TDLM#10 @maximelenormand
It seems that you are using an old version of write_delim in which the "progress" argument is not available. To prevent this I added a minimal version requirement for readr in the DESCRIPTION.
