### 2023-06-05 EpiVec/TDLM#6 @maximelenormand
There are indeed two java folders, one at the root of the package (containing the java source file .java) and another one located in inst containing the compiled files (.class and .jar) used to run the main TDLM's functions. The first one is not used while installing or running the package but it was a CRAN requirement.
