### 2023-06-05 816c990
Add GitHub Actions badge in README (#5)
