### 2023-06-05 EpiVec/TDLM#5 @maximelenormand
I added a GitHub Actions badge to the repo. This is a R-CMD-check performed an the latest version of windows, mac and ubuntu. What do you mean by "build info for a windows or ubuntu build"?
