### 2023-06-05 #5434 @maximelenormand
@crvernon  @kanishkan91. Thank you for accepting to review my package. I started to address @kanishkan91's concerns.

1. The package contains several controls to check the arguments and inputs format. All the functions are documented and automated tested in the examples and in the vignette during the R-CMD-check performed with GitHub Actions on windows, mac and ubuntu. Based on that, the "test process" seems to be between OK and Good according to JOSS Review criteria (https://joss.readthedocs.io/en/latest/review_criteria.html). I am OK to add more automated tests and a badge to rigorously quantify the amount of codes covered by the tests but this may take time.

2. It seems that you are using old versions of the R packages needed to run and install TDLM. I followed the CRAN recommendation by not specifying exact versions and not specifying unnecessarily conservative and inconvenient minimal versions. I will try to go through the imports one by one to identify a minimal version.

3. The "process" argument was added in version 2.0.0 of readr (published in July 2021). I added a minimal version for readr (>= 2.0.0) in the description file. It is not easy to identify the relevant (and not too conservative) minimal version for the six TDLM's dependencies but as mentioned in 2. I will do my best to identify a minimal version for each of them.

4. I am using GitHub actions to perform R-CMD-check on windows, mac and ubuntu. I followed your suggestion and added a badge to the repo.
