### 2023-06-05 EpiVec/TDLM#8 @maximelenormand
I followed the CRAN recommendation by not specifying exact versions and not specifying unnecessarily conservative and inconvenient minimal versions. I will try to go through the imports one by one to identify a relevant and not too conservative minimal version.
