### 2023-06-05 EpiVec/TDLM#3 @maximelenormand
The package contains several controls to check the arguments and inputs format. All the functions are documented and automated tested in the examples and in the vignette during the R-CMD-check performed with GitHub Actions on windows, mac and ubuntu. Based on that, the "test process" seems to be between OK and Good according to JOSS Review criteria (https://joss.readthedocs.io/en/latest/review_criteria.html). I am OK to add more automated tests and a badge to rigorously quantify the amount of codes covered by the tests but this may take time.
