### 2023-06-05 EpiVec/TDLM#7 @maximelenormand
TDLM has six dependencies, the package renv is very interesting though, I will have a look at it for the future development of the package. Thank you for the suggestion.
