### 2023-06-05 EpiVec/TDLM#3 @kanishkan91
@maximelenormand Thanks for the explanation  above. However, I think what would be required would be a testing suite developed via testthat for an R package such as this. This for a couple of reasons,

1. At the moment it is not possible to tell what lines are checked when you run the code
2. If there are developments in this package by a third person, the test cases would help trace if something went wrong and where (i.e. in which function) something went wrong.

I agree it may be a bit time consuming to put this together. But what you can do is just add some simple tests that cover the major functions. It could be something as simple as below,

```
test_that("extract opportunities works correctly",{

    expect_error(extract_opportunities(),NA,info=("Failure to run extract opportunities"))

})
```
