### 2023-06-05 EpiVec/TDLM#5 @kanishkan91
@maximelenormand Currently all there is a R-CMD check. There is no windows build or ubuntu build which would ensure that the package builds correctly on those platforms. You would only have to add a couple of yaml files to get this working. Here is an example- https://github.com/JGCRI/ambrosia/tree/master/.github/workflows

Let me know what you think.
