+++ DESCRIPTION
@@ -14,7 +14,7 @@ Depends: R (>= 4.0.0)
 License: GPL-3
 Encoding: UTF-8
 LazyData: true
-LazyDataCompression:xz
+LazyDataCompression: xz
 Imports:
     Ecume,
     mathjaxr,
@@ -26,9 +26,11 @@ RdMacros:
     mathjaxr,
     Rdpack
 Suggests:
-    knitr
+    knitr,
+    testthat (>= 3.0.0)
 VignetteBuilder: knitr
 RoxygenNote: 7.2.3
 BugReports: https://github.com/EpiVec/TDLM/issues
 Roxygen: list(markdown = TRUE)
 URL: https://epivec.github.io/TDLM/
+Config/testthat/edition: 3
