### 2023-06-07 EpiVec/TDLM#5 @kanishkan91
@maximelenormand I think the issue is that you've only set up the CMD checks which is more of a documentation check. You would also need a windows and ubuntu build so that GitHub would build in that environment and make sure the code compiles. Its just adding a couple more yml files like this- https://github.com/JGCRI/ambrosia/tree/master/.github/workflows

Then you can also add badges for the windows and linux builds.
