### 2023-06-08 EpiVec/TDLM#3 @maximelenormand
yes, I followed your recommendation, I am setting automated test with testthat and use covr to generate a report. I will write tests for all functions within the next few days.
