### 2023-06-08 EpiVec/TDLM#5 @maximelenormand
@kanishkan91 I am not sure to see the difference between https://github.com/EpiVec/TDLM/blob/master/.github/workflows/R-CMD-check.yaml and https://github.com/JGCRI/ambrosia/blob/master/.github/workflows/build_linux.yml. I am running rcmdcheck on 5 configurations (mac, windows,ubuntu devel, ubuntu release and ubuntu oldrel) each time I modify the repo. So the check is passing only if the 5 rcmdcheck are passing.
