+++ tests/testthat/test-run_law.R
@@ -1,0 +1,34 @@
+# Preamble code ----------------------------------------------------------------
+data(mass)
+data(distance)
+data(od)
+
+mi=as.numeric(mass[,1])
+names(mi)=rownames(distance)
+
+sij=extract_opportunities(opportunity=mi, distance=distance, check_names=TRUE)
+
+mj=as.numeric(mass[,1])+0.1
+names(mj)=rownames(distance)
+
+dist=distance
+
+# Tests for valid outputs ------------------------------------------------------
+test_that("class TDLM", {
+  
+  res <- run_law(law = "Unif", mass_origin = mi, mass_destination = NULL, distance = NULL, opportunity = NULL, param = 0.1,
+                 check_names = TRUE)
+  
+  expect_identical(class(res)[1], "TDLM")
+  
+})
+
+# Check errors -----------------------------------------------------------------
+test_that("check errors", {
+  
+  expect_error(run_law(law = "Uni", mass_origin = mi, mass_destination = NULL, distance = NULL, opportunity = NULL, param = 0.1,
+                       check_names = TRUE), 
+               "Please choose law among the followings values:
+GravExp, NGravExp, GravPow, NGravPow, Schneider, Rad, ExtRad or Unif")
+  
+})
