### 2023-06-10 EpiVec/TDLM#12 @kanishkan91
JOSS review: Suggested edits to the paper
The paper currently is only half a page and 350 words. While this is not an issue, there are a few edits that I would suggest,

1.  **Add state of the field section-** Could you compare to any other software that can do this? If there are no software, I assume the only way to do this is to write scripts in R python? Why is this software much better than other options?
2.  **Add software structure section-** Could you add a quick summary of the software structure in the paper? Perhaps explicitly link your vignette in the paper?

https://github.com/openjournals/joss-reviews/issues/5434
