### 2023-06-12 057dbf6
Update DESCRIPTION with minimal version for dependencies
