### 2023-06-12 a442b9c
Implementing automated tests
