+++ tests/testthat/test-utils.R
@@ -1,0 +1,278 @@
+# Check controls ---------------------------------------------------------------
+test_that("check controls vectors_positive", {
+  
+  vec = list()
+  
+  vec[[1]]=c(1,1)
+  #vec$vec = c(1.2,NA)
+  #expect_error(controls(args = NULL, vectors = vec, matrices = NULL, type = "vectors_positive"), 
+  #             "NA(s) detected in vec.")
+  
+  vec$vec = c(1.2,"2")
+  expect_error(controls(args = NULL, vectors = vec, matrices = NULL, type = "vectors_positive"), 
+               "vec must be a numeric vector.")
+  
+  vec$vec = c(1.2)
+  expect_error(controls(args = NULL, vectors = vec, matrices = NULL, type = "vectors_positive"), 
+               "vec must contain at least two locations.")
+  
+  vec$vec = c(0,0)
+  expect_error(controls(args = NULL, vectors = vec, matrices = NULL, type = "vectors_positive"), 
+               "vec must contain at least one strictly positive 
+        value.")
+  
+  vec$vec = c(1.2,-1)
+  expect_error(controls(args = NULL, vectors = vec, matrices = NULL, type = "vectors_positive"), 
+               "vec must contain only positive values.")
+})
+
+test_that("check controls vectors_positive", {
+  
+  vec = list()
+  
+  vec[[1]]=c(1,1)
+  
+  vec$vec = c(1,"2")
+  
+  #vec$vec = c(1,NA)
+  #expect_error(controls(args = NULL, vectors = vec, matrices = NULL, type = "vectors_positive_integer"), 
+  #             "NA(s) detected in vec.")
+  
+  expect_error(controls(args = NULL, vectors = vec, matrices = NULL, type = "vectors_positive_integer"), 
+               "vec must be a numeric vector.")
+  
+  vec$vec = c(1.2,1)
+  expect_error(controls(args = NULL, vectors = vec, matrices = NULL, type = "vectors_positive_integer"), 
+               "vec must be a vector of integers.")
+  
+  vec$vec = c(-1,0)
+  expect_error(controls(args = NULL, vectors = vec, matrices = NULL, type = "vectors_positive_integer"), 
+               "vec must contain only positive values.")
+  
+  vec$vec = c(1)
+  expect_error(controls(args = NULL, vectors = vec, matrices = NULL, type = "vectors_positive_integer"), 
+               "vec must contain at least two locations.")
+  
+  vec$vec = c(0,0)
+  expect_error(controls(args = NULL, vectors = vec, matrices = NULL, type = "vectors_positive_integer"), 
+               "vec must contain at least one strictly positive 
+        value.")
+  
+})
+
+test_that("check controls matrices_positive", {
+  
+  mat = list()
+  
+  mat[[1]] = matrix(1,5,5)
+  
+  mat$mat = matrix(1,5,5)
+  #mat$mat[1,1]=NA
+  #expect_error(controls(args = NULL, vectors = NULL, matrices = mat, type = "matrices_positive"), 
+  #             "NA(s) detected in mat.")
+  
+  mat$mat = 1
+  expect_error(controls(args = NULL, vectors = NULL, matrices = mat, type = "matrices_positive"), 
+               "mat must be a matrix.")
+  
+  mat$mat = matrix(1,5,6)
[…]
