+++ NEWS.md
@@ -3,6 +3,8 @@
 This is a list of changes made in the development/github version of the package
 between TDLM 0.1.0 (CRAN release 2023-03-18) and the next CRAN release.
 
+* Added automated test (testthat + covr)
+
 * Fixed bug in `extract_spatial_information()` (@Pachka, #1).
 
 * `extract_spatial_information()` is now optimized and way more efficient 
