### 2023-06-12 EpiVec/TDLM#10 @maximelenormand
I added a minimal version requirement for every dependency in DESCRIPTION.
