### 2023-06-12 EpiVec/TDLM#8 @maximelenormand
I added a minimal version requirement for every dependency in DESCRIPTION.
