### 2023-06-12 EpiVec/TDLM#12 @maximelenormand
Thanks for the suggestion. I will add the two suggested sections and work on the paper in the next few days.
