### 2023-06-12 #5434 @MAnalytics
@maximelenormand @kanishkan91 @crvernon

Please, try and run googpractice::gp() and fix some minor styling issues.
e.g.,
* using '=' instead of '<-'
* try and truncate long code lines, etc.

Then I can resume my review from there.
