### 2023-06-12 98699a0
Fixed some minor styling issues based on googpractice::gp()
