+++ R/data.R
@@ -1,7 +1,7 @@
 #' Population and number of out- and in-commuters by US Kansas county in 2000
 #'
-#' A dataset containing the number of inhabitants, in-commuters and out-commuters
-#' for 105 US Kansas counties in 2000.
+#' A dataset containing the number of inhabitants, in-commuters and 
+#' out-commuters for 105 US Kansas counties in 2000.
 #'
 #' @format A `data.frame` with 105 rows and 3 columns:
 #' \describe{
