+++ R/utils.R
@@ -2,17 +2,16 @@ controls <- function(args = NULL,
                      vectors = NULL,
                      matrices = NULL,
                      type = "vectors_positive") {
-  
   # vectors_positive ###########################################################
   if (type == "vectors_positive") {
     nbv <- length(vectors)
     for (k in 1:nbv) {
       vector <- vectors[[k]]
       namevec <- names(vectors)[k]
-      
+
       if (sum(is.na(vector)) > 0) {
         stop(paste0("NA(s) detected in ", namevec, "."),
-             call. = FALSE
+          call. = FALSE
         )
       }
 
@@ -46,10 +45,10 @@ controls <- function(args = NULL,
     for (k in 1:nbv) {
       vector <- vectors[[k]]
       namevec <- names(vectors)[k]
-      
+
       if (sum(is.na(vector)) > 0) {
         stop(paste0("NA(s) detected in ", namevec, "."),
-             call. = FALSE
+          call. = FALSE
         )
       }
 
@@ -83,7 +82,7 @@ controls <- function(args = NULL,
           call. = FALSE
         )
       }
-      
+
       if (sum(vector < 0) > 0) {
         stop(paste0(namevec, " must contain only positive values."),
           call. = FALSE
@@ -98,10 +97,10 @@ controls <- function(args = NULL,
     for (k in 1:nbm) {
       matrix <- matrices[[k]]
       namemat <- names(matrices)[k]
-      
+
       if (sum(is.na(matrix)) > 0) {
         stop(paste0("NA(s) detected in ", namemat, "."),
-             call. = FALSE
+          call. = FALSE
         )
       }
 
@@ -133,11 +132,11 @@ controls <- function(args = NULL,
           call. = FALSE
         )
       }
-      #if (sum(diag(matrix)) > 0) {
+      # if (sum(diag(matrix)) > 0) {
       #  stop(paste0("The diagonal of ", namemat, " must be null."),
       #    call. = FALSE
       #  )
-      #}
+      # }
     }
   }
 
@@ -306,7 +305,8 @@ controls <- function(args = NULL,
     # names matrices
     test <- NULL
     for (k in 1:nbm) {
-      test <- c(test, sum(rownames(matrices[[k]]) == colnames(matrices[[k]])) == n)
+      test <- c(test, 
+                sum(rownames(matrices[[k]]) == colnames(matrices[[k]])) == n)
     }
     if (sum(test) < nbm) {
       mess <- "Different rownames and colnames in:\n"
@@ -326,7 +326,8 @@ controls <- function(args = NULL,
       test <- NULL
       for (k1 in 1:(nbm - 1)) {
         for (k2 in (k1 + 1):nbm) {
-          testk1k2 <- (sum(rownames(matrices[[k1]]) == rownames(matrices[[k2]])) == n)
+          testk1k2 <- 
+            (sum(rownames(matrices[[k1]]) == rownames(matrices[[k2]])) == n)
           test <- rbind(test, data.frame(from = k1, to = k2, test = testk1k2))
         }
       }
@@ -395,7 +396,8 @@ controls <- function(args = NULL,
     # names matrices
     test <- NULL
     for (k in 1:nbm) {
-      test <- c(test, sum(rownames(matrices[[k]]) == colnames(matrices[[k]])) == n)
+      test <- c(test, 
+                sum(rownames(matrices[[k]]) == colnames(matrices[[k]])) == n)
[…]
