### 2023-06-13 #5434 @maximelenormand
@MAnalytics @crvernon Thank you for accepting to review my package. I run goodpractice::gp() and fixed some styling issues.
