### 2023-07-04 #5434 @maximelenormand
@MAnalytics @crvernon Could you please elaborate on the difficulties that you encountered while installing the package? I installed it on several virtual windows machine and also asked different colleagues to do it (both the CRAN and the GitHub versions) and I've never encountered any problem.
