### 2023-07-06 #5434 @MAnalytics
As for the installation, apologies it was my PC acting out!
This is a very nice package and I think would be useful in transportation research!
No major issues (separate from the ones already highlighted by @kanishkan91) are identified! But would like to emphasise two minor revisions:
 * Provide a section in the paper describing the software structure, or the main functions of the software.
 * Describe how the software compare with the existing software, and if none exist, this should be emphasized as a major contribution of the paper.
@crvernon I will be happy to recommend the paper for publication as soon as these two issues are addressed.

MA.
