+++ paper.md
@@ -7,7 +7,7 @@ tags:
 - Commuting networks
 - Gravity model
 - Radiation model
-date: "17 March 2023"
+date: "5 July 2023"
 output: pdf_document
 authors:
 - name: "Maxime Lenormand"
@@ -22,17 +22,18 @@ affiliations:
 # Summary
 
 Spatial interaction models are widely used to estimate and explain spatial
-interactions between locations. These models are usually based on the
-characteristics of the locations and the way they are spatially distributed. 
-Interactions between locations can take several forms, population movements, 
-widely studied in geography, transportation research and urban planning is one 
-of them. The flows of individuals between locations is usually represented by a
-trip table better known as Origin-Destination (OD) matrix 
-[@Lenormand2016;@Barbosa2018]. The estimation of OD matrices is part of the 
-four-step travel model in transportation research. It corresponds to the second
-step, called trip distribution, the aim of which is to match the trip origins 
-with the trip destinations using a spatial interaction model commonly referred 
-to as trip distribution model in the four-step travel framework. 
+interactions between geographical areas or locations. These models are usually 
+based on the characteristics of the locations and the way they are 
+spatially distributed. Interactions between locations can take several forms, 
+population movements, widely studied in geography, transportation research and 
+urban planning is one of them. The flows of individuals between locations is 
+usually represented by a trip table better known as Origin-Destination (OD) 
+matrix [@Lenormand2016;@Barbosa2018]. The estimation of OD matrices is part of 
+the four-step travel model in transportation research. It corresponds to the 
+second step, called trip distribution, the aim of which is to match the trip 
+origins with the trip destinations using a spatial interaction model 
+commonly referred to as trip distribution model in the four-step travel 
+framework. 
 
 In order to facilitate the use and comparison of trip distribution models, and 
 more generally spatial distribution models, we present **TDML** an R package 
@@ -41,17 +42,68 @@ trip distribution laws and models as described in @Lenormand2016.
 
 # Statement of need
 
-Trip distribution model are generally composed of two mechanisms, one mechanism
+Trip distribution models are generally composed of two mechanisms, one mechanism
 based on a 'law' to estimate the probability that an individual move from one 
 location to another and a second mechanism based on a 'model' used to estimate 
 the number of individuals moving from one location to another. These two 
 mechanisms are rarely dissociated which could lead to methodological flaws when 
 comparing different laws and/or models [@Lenormand2012;@Simini2012;@Masucci2013;
-@Yang2014].
+@Yang2014]. This is particularly important when we compare the two historical
+approaches - gravity and intervening opportunities - for the estimation of 
+trip distribution. 
 
+We identified several R packages providing an implementation of spatial 
+interaction models that can be used to estimate trip distributions. 
+The **gravity** package [@Wolwer2018], the **spflow** 
+package [@Dargel2021], the **mobility** [@Giles2021] and the **simodels** 
+package [@Lovelace2023]. The **gravity**, **spflow** and **mobility** packages 
+are based on statistical models and have not been designed to compare gravity 
+and intervening opportunities laws and constrained models independently. 
+Although the package structure and functionality of **simodels** are very 
+different from that of **TDLM**, it offers the possibility to compare trip 
+distribution laws and models independently. **simodels** proposes 
+an interesting approach by not defining (nor encouraging the use of) any 
+particular trip distribution laws, but this also makes the comparison 
+between gravity or intervening opportunities laws complicated for non-expert 
+users. Furthermore, **simodels** does not offer the possibility at this stage to
+use the doubly-constrained model nor any function to systematically compare 
+observed and simulated OD matrices.
+ 
 To overcome these limitations, the **TDML** R package is based on a two-step 
[…]
