+++ paper.bib
@@ -59,4 +59,39 @@
   issue = {2}
 }
 
+@Article{Wolwer2018,
+  author  = {W{\"o}lwer, A.-L. and Burgard, J. P. and Kunst, J. and Vargas, M.},
+  journal = {Journal of Open Source Software},
+  title   = {Gravity: Estimation Methods for Gravity Models in R},
+  year    = {2018},
+  number  = {31},
+  pages   = {1038},
+  volume  = {3},
+}
+
+@Manual{Lovelace2023,
+  title  = {simodels: Flexible Framework for Developing Spatial Interaction Models},
+  author = {Lovelace, R. and Nowosad, J.},
+  note   = {https://github.com/robinlovelace/simodels, https://robinlovelace.github.io/simodels/},
+  year   = {2023},
+}
+
+@Manual{Dargel2021,
+  title  = {spflow: Spatial Econometric Interaction Models},
+  author = {Dargel, L. and Laurent, T.},
+  note   = {R package version 0.1.0},
+  year   = {2021},
+  url    = {https://CRAN.R-project.org/package=spflow},
+}
+
+@Software{Giles2021,
+  author    = {Giles, J.},
+  doi       = {10.5281/zenodo.5171373},
+  publisher = {Zenodo},
+  title     = {{COVID-19-Mobility-Data-Network/mobility: mobility version 0.6.4}},
+  url       = {https://doi.org/10.5281/zenodo.5171373},
+  version   = {v0.6.4},
+  year      = {2021},
+}
+
 @Comment{jabref-meta: databaseType:bibtex;}
