+++ paper.md
@@ -64,10 +64,10 @@ different from that of **TDLM**, it offers the possibility to compare trip
 distribution laws and models independently. **simodels** proposes 
 an interesting approach by not defining (nor encouraging the use of) any 
 particular trip distribution laws, but this also makes the comparison 
-between gravity or intervening opportunities laws complicated for non-expert 
-users. Furthermore, **simodels** does not offer the possibility at this stage to
-use the doubly-constrained model nor any function to systematically compare 
-observed and simulated OD matrices.
+between gravity or intervening opportunities laws more complicated for 
+non-expert users. Furthermore, **simodels** does not offer the possibility at 
+this stage to use the doubly-constrained model or any functionality to 
+systematically compare observed and simulated OD matrices.
  
 To overcome these limitations, the **TDML** R package is based on a two-step 
 approach to generate mobility flows by separating the trip distribution law, 
