### 2023-07-10 #5434 @maximelenormand
@MAnalytics @kanishkan91 @crvernon Thanks a lot for your reviews. I've just updated the paper by adding a package "Functionality" section describing the main functions of the package. I also added a description and limitations of existing R packages in the "Statement of need" section. Thanks again for the suggestions.
