### 2023-07-25 5d047d1
Update paper based on @crvernon final comments
