+++ paper.bib
@@ -1,10 +1,11 @@
 @Article{Barbosa2018,
-  Title                    = {Human mobility: Models and applications},
-  Author                   = {Barbosa, H. and Barthelemy, M. and Ghoshal, G. and James, C. R. and Lenormand, M. and Louail, T. and Menezes, R. and Ramasco, J. J. and Simini, F. and Tomasini, M.},
-  Journal                  = {Physics Reports},
-  Year                     = {2018},
-  Pages                    = {1-74},
-  Volume                   = {734}
+  author  = {Barbosa, H. and Barthelemy, M. and Ghoshal, G. and James, C. R. and Lenormand, M. and Louail, T. and Menezes, R. and Ramasco, J. J. and Simini, F. and Tomasini, M.},
+  journal = {Physics Reports},
+  title   = {Human mobility: Models and applications},
+  year    = {2018},
+  pages   = {1-74},
+  volume  = {734},
+  doi     = {10.1016/j.physrep.2018.01.001},
 }
 
 @Article{Lenormand2016,
@@ -14,49 +15,51 @@
   year    = {2016},
   pages   = {158-169},
   volume  = {51},
+  doi     = {10.1016/j.jtrangeo.2015.12.008},
 }
 
-@ARTICLE{Simini2012,
-  author = {Simini, F. and Gonz{\'{a}}lez, M. C. and Maritan, A. and Barabasi,
-	A.-L.},
-  title = {A universal model for mobility and migration patterns},
+@Article{Simini2012,
+  author  = {Simini, F. and Gonz{\'{a}}lez, M. C. and Maritan, A. and Barabasi, A.-L.},
   journal = {Nature},
-  year = {2012},
-  volume = {484},
-  pages = {96-100}
+  title   = {A universal model for mobility and migration patterns},
+  year    = {2012},
+  pages   = {96-100},
+  volume  = {484},
+  doi     = {10.1038/nature10856},
 }
 
-@ARTICLE{Yang2014,
-  author = {Yang, Y. and Herrera, C. and Eagle, N. and Gonz{\'{a}}lez, M. C.},
-  title = {Limits of Predictability in Commuting Flows in the Absence of Data
-	for Calibration},
+@Article{Yang2014,
+  author  = {Yang, Y. and Herrera, C. and Eagle, N. and Gonz{\'{a}}lez, M. C.},
   journal = {Scientific Reports},
-  year = {2014},
-  volume = {4},
-  pages = {5662},
-  number = {5662}
+  title   = {Limits of Predictability in Commuting Flows in the Absence of Data for Calibration},
+  year    = {2014},
+  number  = {5662},
+  pages   = {5662},
+  volume  = {4},
+  doi     = {10.1038/srep05662},
 }
 
-@ARTICLE{Lenormand2012,
-  author = {Lenormand, M. and Huet, S. and Gargiulo, F. and Deffuant, G.},
-  title = {A {U}niversal {M}odel of {C}ommuting {N}etworks},
-  journal = {PLoS ONE},
-  year = {2012},
-  volume = {7},
-  pages = {e45985},
-  owner = {maxime.lenormand},
-  timestamp = {2012.03.23}
+@Article{Lenormand2012,
+  author    = {Lenormand, M. and Huet, S. and Gargiulo, F. and Deffuant, G.},
+  journal   = {PLoS ONE},
+  title     = {A {U}niversal {M}odel of {C}ommuting {N}etworks},
+  year      = {2012},
+  pages     = {e45985},
+  volume    = {7},
+  doi       = {10.1371/journal.pone.0045985},
+  owner     = {maxime.lenormand},
+  timestamp = {2012.03.23},
 }
 
-@ARTICLE{Masucci2013,
[…]
