+++ paper.md
@@ -7,7 +7,7 @@ tags:
 - Commuting networks
 - Gravity model
 - Radiation model
-date: "5 July 2023"
+date: "25 July 2023"
 output: pdf_document
 authors:
 - name: "Maxime Lenormand"
@@ -24,10 +24,11 @@ affiliations:
 Spatial interaction models are widely used to estimate and explain spatial
 interactions between geographical areas or locations. These models are usually 
 based on the characteristics of the locations and the way they are 
-spatially distributed. Interactions between locations can take several forms, 
-population movements, widely studied in geography, transportation research and 
-urban planning is one of them. The flows of individuals between locations is 
-usually represented by a trip table better known as Origin-Destination (OD) 
+spatially distributed. Several forms of interaction can occur between locations,
+one of which is population movements. This particular type of interaction is 
+widely examined in the fields of geography, transportation research, and 
+urban planning. The flows of individuals between locations is 
+usually represented by a trip table better known as an Origin-Destination (OD) 
 matrix [@Lenormand2016;@Barbosa2018]. The estimation of OD matrices is part of 
 the four-step travel model in transportation research. It corresponds to the 
 second step, called trip distribution, the aim of which is to match the trip 
@@ -37,16 +38,16 @@ framework.
 
 In order to facilitate the use and comparison of trip distribution models, and 
 more generally spatial distribution models, we present **TDML** an R package 
-proposing a set of easy-to-use functions to rigorously and fairly compare 
+providing a set of easy-to-use functions to rigorously and fairly compare 
 trip distribution laws and models as described in @Lenormand2016.
 
 # Statement of need
 
 Trip distribution models are generally composed of two mechanisms, one mechanism
-based on a 'law' to estimate the probability that an individual move from one 
+based on a 'law' to estimate the probability that an individual moves from one 
 location to another and a second mechanism based on a 'model' used to estimate 
 the number of individuals moving from one location to another. These two 
-mechanisms are rarely dissociated which could lead to methodological flaws when 
+mechanisms are rarely disentangled which could lead to methodological flaws when 
 comparing different laws and/or models [@Lenormand2012;@Simini2012;@Masucci2013;
 @Yang2014]. This is particularly important when we compare the two historical
 approaches - gravity and intervening opportunities - for the estimation of 
@@ -69,16 +70,15 @@ of trip distribution laws more complicated for non-expert users. Furthermore,
 observed and simulated OD matrices.
  
 To overcome these limitations, the **TDML** R package is based on a two-step 
-approach to generate mobility flows by separating the trip distribution law, 
-gravity or intervening opportunities, from the modeling approach used to 
-generate the flows from this law. 
+approach to generate mobility flows by separating the trip distribution law 
+from the modeling approach used to generate the flows from this law. 
 
 # Functionality
 
 **TDLM** is available on [CRAN](https://cran.r-project.org/package=TDLM) and 
 [GitHub](https://github.com/EpiVec/TDLM/). The **TDLM**'s website includes a [tutorial](https://epivec.github.io/TDLM/articles/TDLM.html) 
 describing the functions of this package with an illustrative example based on 
-commuting data from US Kansas in 2000.
+commuting data from Kansas in the United States in 2000.
 
 **TDLM** features four main functions for generating OD matrices
 based on a wide range of trip distribution laws and models and for evaluating
