### 2023-07-25 #5434 @crvernon
Thank you @maximelenormand !

I am going to move forward with getting this paper ready for acceptance.

So here is what we have left to do:

- [ ] Conduct a GitHub release of the current reviewed version of the software you now have on the main and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)

- [ ] Please list the DOI of the archived version here
