### 2023-08-11 c6a673a
small changes in paper
