### 2023-08-11 #5434 @danielskatz
Hi @maximelenormand - I'm the track editor who will handle the final processing of your submission. I've proofread the draft paper, and have suggested some changes in https://github.com/EpiVec/TDLM/pull/13 Please merge this, or let me know what you disagree with, then we can continue to acceptance and publication.
