### 2023-08-11 #5434 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
