### 2023-08-11 #5434 @danielskatz
Congratulations to @maximelenormand (Maxime Lenormand) on your publication!!

JOSS is volunteer-run, and is completely dependent on your efforts
